+++ pydidas/workflow/workflow_node.py
@@ -30,9 +30,11 @@ __all__ = ["WorkflowNode"]
 
 from copy import deepcopy
 from numbers import Integral
+from typing import Union
 
+from ..core import Dataset
+from ..core.utils import LOGGING_LEVEL, TimerSaveRuntime, pydidas_logger
 from ..plugins import BasePlugin
-from ..core.utils import pydidas_logger, LOGGING_LEVEL, TimerSaveRuntime
 from .generic_node import GenericNode
 
 
@@ -42,46 +44,38 @@ logger.setLevel(LOGGING_LEVEL)
 
 class WorkflowNode(GenericNode):
     """
-    The WorkflowNode subclass of the GenericNode has an added plugin attribute
-    to allow it to execute plugins, either individually or the full chain.
+    A subclassed GenericNode wit han added plugin attribute.
+
+    The WorkflowNode allows executing plugins individually or in a full workflow
+    chain through the WorkflowTree.
     """
 
     kwargs_for_copy_creation = ["plugin", "_result_shape"]
 
-    def __init__(self, **kwargs):
-        kwargs = self.__preprocess_kwargs(**kwargs)
-        super().__init__(**kwargs)
+    def __init__(self, **kwargs: dict):
+        self.__preprocess_kwargs(kwargs)
+        GenericNode.__init__(self, **kwargs)
         self.__confirm_plugin_existance_and_type()
-        self.__process_node_id()
-        self.plugin.node_id = self._node_id
+        self.node_id = self.__tmp_node_id
         self.results = None
         self.result_kws = None
         self._result_shape = None
         self.runtime = -1
 
-    def __preprocess_kwargs(self, **kwargs):
+    def __preprocess_kwargs(self, kwargs: dict):
         """
-        Preprocess the keyword arguments and store and remove the node_id key
-        from them.
+        Store and remove the node_id key from the calling kwargs.
 
         The node ID key must be set after the plugin has been stored to
         force syncing of Plugin node ID and WorkflowNode node ID.
 
         Parameters
         ----------
-        **kwargs : dict
-            The calling keyword arguments.
-
-        Returns
-        -------
         kwargs : dict
-            The calling keyword arguments (minus the node_id key).
+            The calling keyword arguments from init.
         """
         self.plugin = None
-        self.__tmp_node_id = kwargs.get("node_id", None)
-        if "node_id" in kwargs:
-            del kwargs["node_id"]
-        return kwargs
+        self.__tmp_node_id = kwargs.pop("node_id", None)
 
     def __confirm_plugin_existance_and_type(self):
         """
@@ -96,23 +90,14 @@ class WorkflowNode(GenericNode):
         """
         if self.plugin is None:
             raise KeyError(
-                "No plugin has been supplied for the WorkflowNode."
-                " Node has not been created."
+                "No plugin has been supplied for the WorkflowNode. "
+                "The node has not been created."
             )
         if not isinstance(self.plugin, BasePlugin):
             raise TypeError("Plugin must be an instance of BasePlugin (or subclass).")
 
-    def __process_node_id(self):
-        """
-        Process the node ID and set if for both the WorkflowNode and the
-        Plugin.
-        """
-        if self.__tmp_node_id is not None:
-            self.node_id = self.__tmp_node_id
-        del self.__tmp_node_id
-
     @property
-    def node_id(self):
+    def node_id(self) -> Union[int, None]:
         """
         Get the node_id.
 
@@ -127,7 +112,7 @@ class WorkflowNode(GenericNode):
         return self._node_id
 
     @node_id.setter
-    def node_id(self, new_id):
+    def node_id(self, new_id: Union[None, int]):
         """
         Set the node_id.
 
@@ -141,15 +126,14 @@ class WorkflowNode(GenericNode):
         TypeError
             If the type of the new ID is not int or None.
         """
-        if new_id is None or isinstance(new_id, Integral):
-            self._node_id = new_id
-            self.plugin.node_id = new_id
-            return
-        raise TypeError(
-            "The new node_id is not of a correct type and has not been set."
-        )
+        if not (new_id is None or isinstance(new_id, Integral)):
+            raise TypeError(
+                "The new node_id is not of a correct type and has not been set."
+            )
+        self._node_id = new_id
+        self.plugin.node_id = new_id
 
-    def consistency_check(self):
+    def consistency_check(self) -> bool:
         """
         Property to determine if the data is consistent.
 
@@ -169,21 +153,22 @@ class WorkflowNode(GenericNode):
 
     def prepare_execution(self):
         """
-        Prepare the execution of the plugin chain by calling the pre_execute
-        methods of all plugins.
+        Prepare the execution of the plugin chain.
+
+        This method recursively calls the pre_execute methods of all (child) plugins.
         """
         self.plugin.pre_execute()
         for _child in self._children:
             _child.prepare_execution()
 
-    def execute_plugin(self, *args, **kwargs):
+    def execute_plugin(self, arg: Union[Dataset, int], **kwargs: dict):
         """
         Execute the plugin associated with the node.
 
         Parameters
         ----------
-        *args : tuple
-            Any arguments which need to be passed to the plugin.
+        arg : Union[Dataset, int]
+            The argument which need to be passed to the plugin.
         **kwargs : dict
             Any keyword arguments which need to be passed to the plugin.
 
@@ -196,12 +181,12 @@ class WorkflowNode(GenericNode):
         """
         with TimerSaveRuntime() as _runtime:
             if kwargs.get("store_input_data", False):
-                self.plugin.store_input_data_copy(*args, **kwargs)
-            _results = self.plugin.execute(*args, **kwargs)
+                self.plugin.store_input_data_copy(arg, **kwargs)
+            _results = self.plugin.execute(arg, **kwargs)
         self.runtime = _runtime()
         return _results
 
-    def execute_plugin_chain(self, arg, **kwargs):
+    def execute_plugin_chain(self, arg: Union[Dataset, int], **kwargs: dict):
         """
         Execute the full plugin chain recursively.
 
@@ -212,8 +197,8 @@ class WorkflowNode(GenericNode):
 
         Parameters
         ----------
-        *args : tuple
-            Any arguments which need to be passed to the plugin.
+        arg : Union[Dataset, int]
+            The argument which need to be passed to the plugin.
         **kwargs : dict
             Any keyword arguments which need to be passed to the plugin.
         """
@@ -226,14 +211,21 @@ class WorkflowNode(GenericNode):
             or self.plugin.get_param_value("keep_results")
             or kwargs.get("force_store_results", False)
         ) and self.plugin.output_data_dim is not None:
-            self.results = res
+            self.results = deepcopy(res)
             self.result_kws = reskws
         self.runtime = _runtime()
         for _child in self._children:
-            _child.execute_plugin_chain(res, **self._get_deep_copy_of_kwargs(reskws))
+            if len(self._children) > 1:
+                _child.execute_plugin_chain(
+                    deepcopy(res), **self._get_deep_copy_of_kwargs(reskws)
+                )
+            else:
+                _child.execute_plugin_chain(
+                    res, **self._get_deep_copy_of_kwargs(reskws)
+                )
 
     @staticmethod
-    def _get_deep_copy_of_kwargs(kwargs):
+    def _get_deep_copy_of_kwargs(kwargs: dict) -> dict:
         """
         Get a recursive deep copy of the kwargs.
 
@@ -252,7 +244,7 @@ class WorkflowNode(GenericNode):
         }
         return _kwargs_copy
 
-    def dump(self):
+    def dump(self) -> dict:
         """
         Dump the node to a saveable format.
 
@@ -267,31 +259,28 @@ class WorkflowNode(GenericNode):
         dict
             The dict with all required information about the node.
         """
-        _parent = None if self.parent is None else self.parent._node_id
-        _children = [child._node_id for child in self._children]
-        _rep = dict(
-            node_id=self.node_id,
-            parent=_parent,
-            children=_children,
-            plugin_class=self.plugin.__class__.__name__,
-            plugin_params=[
+        _parent = None if self.parent is None else self.parent.node_id
+        _children = [child.node_id for child in self._children]
+        return {
+            "node_id": self.node_id,
+            "parent": _parent,
+            "children": _children,
+            "plugin_class": self.plugin.__class__.__name__,
+            "plugin_params": [
                 p.export_refkey_and_value() for p in self.plugin.params.values()
             ],
-        )
-        return _rep
+        }
 
     def propagate_shapes_and_global_config(self):
         """
-        Calculate the Plugin's shape of the results and push this to the node's
-        children.
+        Calculate the Plugin's result shape results and push it to the node's children.
         """
         self.update_plugin_result_data_shape()
         self.propagate_to_children()
 
     def update_plugin_result_data_shape(self):
         """
-        Update the result shape from the Plugin's input image shape and legacy
-        operations.
+        Update the result shape from the Plugin's input shape and legacy operations.
         """
         self.plugin.update_legacy_image_ops_with_this_plugin()
         self.plugin.calculate_result_shape()
@@ -321,7 +310,7 @@ class WorkflowNode(GenericNode):
         """
         return self._result_shape
 
-    def __hash__(self):
+    def __hash__(self) -> int:
         """
         Get the hash value for the WorkflowNode.