+++ pydidas/workflow/workflow_tree_io/workflow_tree_io_yaml.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2023, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -12,22 +14,30 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
+
 """
 Module with the WorkflowTreeIoYaml class to import/export the WorkflowTree to yaml
 files.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
-__status__ = "Development"
+__status__ = "Production"
 __all__ = ["WorkflowTreeIoYaml"]
+
+
+from pathlib import Path
+from typing import NewType, Union
 
 import yaml
 
 from ...core.constants import YAML_EXTENSIONS
 from .workflow_tree_io_base import WorkflowTreeIoBase
+
+
+WorkflowTree = NewType("WWorkflowTree", type)
 
 
 class WorkflowTreeIoYaml(WorkflowTreeIoBase):
@@ -39,7 +49,9 @@ class WorkflowTreeIoYaml(WorkflowTreeIoBase):
     format_name = "YAML"
 
     @classmethod
-    def export_to_file(cls, filename, tree, **kwargs):
+    def export_to_file(
+        cls, filename: Union[Path, str], tree: WorkflowTree, **kwargs: dict
+    ):
         """
         Write the content to a file.
 
@@ -49,8 +61,8 @@ class WorkflowTreeIoYaml(WorkflowTreeIoBase):
         ----------
         filename : str
             The filename of the file to be written.
-        content : type
-            The content in any format.
+        tree : WorkflowTree
+            The workflow tree instance.
         """
         cls.check_for_existing_file(filename, **kwargs)
         _dump = [node.dump() for node in tree.nodes.values()]
@@ -58,15 +70,15 @@ class WorkflowTreeIoYaml(WorkflowTreeIoBase):
             yaml.safe_dump(_dump, _file)
 
     @classmethod
-    def import_from_file(cls, filename):
+    def import_from_file(cls, filename: Union[Path, str]) -> WorkflowTree:
         """
-        Restore the content from a file
+        Restore the content from a file.
 
         This method needs to be implemented by the concrete subclass.
 
         Parameters
         ----------
-        filename : str
+        filename : Union[Path, str]
             The filename of the file to be written.
 
         Returns